caching-llm-responses/requirements.txt
sentence-transformers>=2.0.0
faiss-cpu>=1.7.0
numpy>=1.24.0
